Fix other data sources, mypy issues, ACS aggregation, and set memory to 4GB
