+++ config/data_sources/AGG_acs_population.sql
@@ -38,19 +38,19 @@ CREATE TEMP FUNCTION getStaggeredDecadeAgeBuckets(x ANY TYPE) AS (
 CREATE OR REPLACE TABLE acs_population.by_sex_age_state AS
 SELECT state_fips, state_name, sex, getDecadeAgeBucket(age) AS age, SUM(population) AS population
 FROM `acs_population.by_sex_age_race_state_std`
-WHERE race_and_ethnicity = "Total"
+WHERE race_category_id = "TOTAL"
 GROUP BY state_fips, state_name, sex, age
 ORDER BY state_fips, state_name, sex, age;
 
 CREATE OR REPLACE TABLE acs_population.by_sex_age_county AS
 SELECT state_fips, county_fips, county_name, sex, getDecadeAgeBucket(age) AS age, SUM(population) AS population
 FROM `acs_population.by_sex_age_race_county_std`
-WHERE race_and_ethnicity = "Total"
+WHERE race_category_id = "TOTAL"
 GROUP BY state_fips, county_fips, county_name, sex, age
 ORDER BY state_fips, county_fips, county_name, sex, age;
 
 -- We can base further aggregations on the above tables. No need to filter to
--- race_and_ethnicity = "Total" since the above tables have already done that.
+-- race_category_id = "TOTAL" since the above tables have already done that.
 CREATE OR REPLACE TABLE acs_population.by_age_state AS
 SELECT state_fips, state_name, age, SUM(population) AS population
 FROM `acs_population.by_sex_age_state`
@@ -67,41 +67,53 @@ ORDER BY state_fips, county_fips, county_name, age;
 -- These tables use staggered decade age buckets due to limitations with ACS
 -- age bucket availability.
 CREATE OR REPLACE TABLE acs_population.by_sex_age_race_state_staggered_buckets AS
-SELECT state_fips, state_name, sex, getStaggeredDecadeAgeBuckets(age) AS age, race_and_ethnicity, SUM(population) AS population
+SELECT
+  state_fips, state_name, sex, getStaggeredDecadeAgeBuckets(age) AS age,
+  race_category_id, race, race_includes_hispanic, race_and_ethnicity,
+  SUM(population) AS population
 FROM `acs_population.by_sex_age_race_state_std`
-GROUP BY state_fips, state_name, sex, age, race_and_ethnicity
-ORDER BY state_fips, state_name, sex, age, race_and_ethnicity;
+GROUP BY state_fips, state_name, sex, age, race_category_id, race, race_includes_hispanic, race_and_ethnicity
+ORDER BY state_fips, state_name, sex, age, race_category_id, race, race_includes_hispanic, race_and_ethnicity;
 
 CREATE OR REPLACE TABLE acs_population.by_sex_age_race_county_staggered_buckets AS
-SELECT state_fips, county_fips, county_name, sex, getStaggeredDecadeAgeBuckets(age) AS age, race_and_ethnicity, SUM(population) AS population
+SELECT
+  state_fips, county_fips, county_name, sex, getStaggeredDecadeAgeBuckets(age) AS age,
+  race_category_id, race, race_includes_hispanic, race_and_ethnicity,
+  SUM(population) AS population
 FROM `acs_population.by_sex_age_race_county_std`
-GROUP BY state_fips, county_fips, county_name, sex, age, race_and_ethnicity
-ORDER BY state_fips, county_fips, county_name, sex, age, race_and_ethnicity;
+GROUP BY state_fips, county_fips, county_name, sex, age, race_category_id, race, race_includes_hispanic, race_and_ethnicity
+ORDER BY state_fips, county_fips, county_name, sex, age, race_category_id, race, race_includes_hispanic, race_and_ethnicity;
 
 -- We can base further aggregations on the above tables.
 CREATE OR REPLACE TABLE acs_population.by_age_race_state_staggered_buckets AS
-SELECT state_fips, state_name, age, race_and_ethnicity, SUM(population) AS population
+SELECT
+  state_fips, state_name, age,
+  race_category_id, race, race_includes_hispanic, race_and_ethnicity,
+  SUM(population) AS population
 FROM `acs_population.by_sex_age_race_state_staggered_buckets`
-GROUP BY state_fips, state_name, age, race_and_ethnicity
-ORDER BY state_fips, state_name, age, race_and_ethnicity;
+GROUP BY state_fips, state_name, age, race_category_id, race, race_includes_hispanic, race_and_ethnicity
+ORDER BY state_fips, state_name, age, race_category_id, race, race_includes_hispanic, race_and_ethnicity;
 
 CREATE OR REPLACE TABLE acs_population.by_age_race_county_staggered_buckets AS
-SELECT state_fips, county_fips, county_name, age, race_and_ethnicity, SUM(population) AS population
+SELECT
+  state_fips, county_fips, county_name, age
+  race_category_id, race, race_includes_hispanic, race_and_ethnicity,
+  SUM(population) AS population
 FROM `acs_population.by_sex_age_race_county_staggered_buckets`
-GROUP BY state_fips, county_fips, county_name, age, race_and_ethnicity
-ORDER BY state_fips, county_fips, county_name, age, race_and_ethnicity;
+GROUP BY state_fips, county_fips, county_name, age, race_category_id, race, race_includes_hispanic, race_and_ethnicity
+ORDER BY state_fips, county_fips, county_name, age, race_category_id, race, race_includes_hispanic, race_and_ethnicity;
 
 
 CREATE OR REPLACE TABLE acs_population.by_sex_state AS
 SELECT state_fips, state_name, sex, SUM(population) AS population
 FROM `acs_population.by_sex_age_race_state_std`
-WHERE race_and_ethnicity = "Total"
+WHERE race_category_id = "TOTAL"
 GROUP BY state_fips, state_name, sex
 ORDER BY state_fips, state_name, sex;
 
 CREATE OR REPLACE TABLE acs_population.by_sex_county AS
 SELECT state_fips, county_fips, county_name, sex, SUM(population) AS population
 FROM `acs_population.by_sex_age_race_county_std`
-WHERE race_and_ethnicity = "Total"
+WHERE race_category_id = "TOTAL"
 GROUP BY state_fips, county_fips, county_name, sex
 ORDER BY state_fips, county_fips, county_name, sex;
