+++ python/datasources/acs_health_insurance.py
@@ -10,7 +10,9 @@ from ingestion.standardized_columns import (STATE_FIPS_COL, COUNTY_FIPS_COL,
                                             RACE_CATEGORY_ID_COL,
                                             WITH_HEALTH_INSURANCE_COL,
                                             WITHOUT_HEALTH_INSURANCE_COL,
-                                            TOTAL_HEALTH_INSURANCE_COL, Race)
+                                            TOTAL_HEALTH_INSURANCE_COL, Race,
+                                            add_race_columns_from_category_id,
+                                            RACE_INCLUDES_HISPANIC_COL)
 from typing import Dict
 
 # TODO pass this in from message data.
@@ -288,6 +290,8 @@ class AcsHealhInsuranceIngestor:
         for table_name, df in self.frames.items():
             # All breakdown columns are strings
             column_types = {c: 'STRING' for c in df.columns}
+            if RACE_INCLUDES_HISPANIC_COL in df.columns:
+                column_types[RACE_INCLUDES_HISPANIC_COL] = 'BOOL'
 
             column_types[WITH_HEALTH_INSURANCE_COL] = 'INT64'
             column_types[WITHOUT_HEALTH_INSURANCE_COL] = 'INT64'
@@ -579,6 +583,9 @@ class AcsHealhInsuranceIngestor:
             WITH_HEALTH_INSURANCE_COL,
             WITHOUT_HEALTH_INSURANCE_COL,
             TOTAL_HEALTH_INSURANCE_COL])
+
+        add_race_columns_from_category_id(self.state_race_frame)
+        add_race_columns_from_category_id(self.county_race_frame)
 
         # Aggregate Frames by Filename
         self.frames = {
